Organize into packages
Different class objects have different purposes, some of which can be grouped.

Let's formalise the groupings by introducing 3 packages.
