+++ src/main/java/Duchess/TaskObjects/Deadline.java
@@ -1,4 +1,4 @@
-package Duchess;
+package Duchess.TaskObjects;
 
 public class Deadline extends Task {
