+++ src/main/java/Duchess/TaskObjects/ToDo.java
@@ -1,4 +1,4 @@
-package Duchess;
+package Duchess.TaskObjects;
 
 public class ToDo extends Task {
